chore(account-lib): code changes post move
- changes required to compile baseCoin in sdk-core
- adds strictPropertyInitialization - false to tsconfig

Ticket: BG-47954
