+++ modules/sdk-core/package.json
@@ -40,6 +40,11 @@
     ]
   },
   "dependencies": {
-    "debug": "3.1.0"
+    "@bitgo/statics": "^6.17.0",
+    "bip32": "^2.0.6",
+    "bitcoinjs-message": "^2.0.0",
+    "bs58": "^2.0.1",
+    "debug": "3.1.0",
+    "tweetnacl": "^1.0.3"
   }
 }
