+++ modules/sdk-core/src/account-lib/baseCoin/baseTransactionBuilderFactory.ts
@@ -1,5 +1,4 @@
 import { BaseCoin as CoinConfig } from '@bitgo/statics';
-import { DeployUtil } from 'casper-js-sdk';
 
 /**
  * Generic transaction builder factory to be extended with coin specific logic.
@@ -29,7 +28,7 @@ export abstract class BaseTransactionBuilderFactory {
   /**
    * Returns a proper builder for the given encoded transaction
    *
-   * @param {Uint8Array | string | DeployUtil.Deploy} raw - Encoded transaction in hex string or Uint8Array format
+   * @param {Uint8Array | string} raw - Encoded transaction in hex string or Uint8Array format
    */
-  public abstract from(raw: Uint8Array | string | DeployUtil.Deploy);
+  public abstract from(raw: Uint8Array | string);
 }
