+++ modules/sdk-core/src/account-lib/baseCoin/index.ts
@@ -7,7 +7,5 @@ export { Ed25519KeyPair } from './ed25519KeyPair';
 export { TransactionType, StakingOperationTypes, AddressFormat } from './enum';
 export { BlsKeyPair } from './blsKeyPair';
 
-import * as Error from './errors';
-import * as Interface from './iface';
-
-export { Error, Interface };
+export * from './errors';
+export * from './iface';
