+++ modules/sdk-core/src/index.ts
@@ -1,3 +1,4 @@
+export * from './account-lib';
 export * from './v2';
 export { BitGoJsError } from './bigojsError';
 import * as common from './common';
